Wired into the Most updated version of the SettingsComposable
Set some constants
Removed references to the legacy version
Changed the location of the package
