+++ app/src/main/java/com/brainwallet/constants/BWConstants.kt
@@ -103,9 +103,9 @@ object BWConstants {
     const val TWITTER_LINK: String = "https://twitter.com/Brainwallet_App"
     const val INSTAGRAM_LINK: String = "https://www.instagram.com/brainwalletapp"
     const val WEB_LINK: String = "https://brainwallet.co"
-    const val SUPPORT_WEB_LINK: String = "https://brainwallet.co/support.html"
+    const val SUPPORT_WEB_LINK: String = "https://www.brainwallet.co/support"
     const val TOS_LINK: String = "https://brainwallet.co/privacy-policy.html"
-    var LINKTREE_URL: String = "https://linktr.ee/brainwallet"
+    const val LINKTREE_URL: String = "https://linktr.ee/brainwallet"
 
     /**
      * API Hosts
