+++ app/src/main/java/com/brainwallet/presenter/activities/BreadActivity.java
@@ -57,7 +57,7 @@ import com.brainwallet.tools.util.ExtensionKt;
 import com.brainwallet.tools.util.Utils;
 import com.brainwallet.ui.BrainwalletActivity;
 import com.brainwallet.ui.screens.settings.SettingsViewModel;
-import com.brainwallet.ui.screens.main.composable.HomeSettingDrawerComposeView;
+import com.brainwallet.ui.screens.settings.settingsrows.HomeSettingDrawerComposeView;
 import com.brainwallet.ui.bentosections.buyreceivebento.receive.ReceiveDialogFragment;
 import com.brainwallet.util.PermissionUtil;
 import com.brainwallet.wallet.BRPeerManager;
@@ -389,7 +389,6 @@ public class BreadActivity extends BRActivity implements BRWalletManager.OnBalan
 
         navigationDrawer = findViewById(R.id.navigationDrawer);
         drawerLayout = findViewById(R.id.drawerLayout);
-        homeSettingDrawerComposeView = findViewById(R.id.homeDrawerComposeView);
         homeSettingDrawerComposeView.observeBus(message -> {
             drawerLayout.close();
             if (SettingsViewModel.LEGACY_EFFECT_ON_LOCK.equals(message.getMessage())) {
