+++ app/src/main/java/com/brainwallet/ui/screens/settings/BentoSettingsButton.kt
@@ -18,6 +18,7 @@ import androidx.compose.ui.tooling.preview.PreviewLightDark
 import androidx.compose.ui.unit.dp
 import com.brainwallet.ui.theme.DesignTheme
 import com.brainwallet.R
+import com.brainwallet.constants.topNavButtonSize
 import com.brainwallet.ui.theme.bentoDarkBorderGradient
 import com.brainwallet.ui.theme.bentoDarkSurfaceGradient
 import com.brainwallet.ui.theme.bentoLightBorderGradient
@@ -33,7 +34,7 @@ fun BentoSettingsButton(
     Box(
         contentAlignment = Alignment.Center,
         modifier = modifier
-            .size(48.dp)
+            .size(topNavButtonSize)
             .clip(CircleShape)
             .background(
                 brush = if (isDarkMode) bentoDarkSurfaceGradient else bentoLightSurfaceGradient,
