+++ app/src/main/java/com/brainwallet/ui/screens/settings/settingsrows/LanguageDetail.kt
@@ -1,12 +1,9 @@
-package com.brainwallet.ui.screens.main.composable.settingsrows
+package com.brainwallet.ui.screens.settings.settingsrows
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
@@ -22,67 +19,54 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.alpha
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import com.brainwallet.R
-import com.brainwallet.data.model.CurrencyEntity
-import com.brainwallet.tools.sqlite.CurrencyDataSource
+import com.brainwallet.data.model.Language
 import com.brainwallet.ui.theme.DesignTheme
 
+// TODO
 @Composable
-fun CurrencyDetail(
+fun LanguageDetail(
     modifier: Modifier = Modifier,
-    selectedCurrency: CurrencyEntity,
-    onFiatSelect: (CurrencyEntity) -> Unit
+    selectedLanguage: Language,
+    onLanguageSelect: (Language) -> Unit
 ) {
-    val context = LocalContext.current
+    val languages = Language.entries.toTypedArray()
 
     // / Layout values
-    val closedHeight = 60
-    val expandedHeight = 100
-    val dividerThickness = 1
+    val expandedHeight = 150
     val unselectedCircleSize = 20
-    val tinyPad = 2
 
     SettingRowItemExpandable(
         modifier = modifier,
-        title = stringResource(R.string.settings_title_currency)
+        title = stringResource(R.string.settings_title_language)
     ) {
-        LazyColumn(
-            modifier = Modifier.height(expandedHeight.dp)
-        ) {
+        LazyColumn(modifier = Modifier.height(expandedHeight.dp)) {
             items(
-                items = CurrencyDataSource.getInstance(context).getAllCurrencies(true)
-            ) { currency ->
+                items = languages
+            ) { language ->
                 ListItem(
                     colors = ListItemDefaults.colors(
                         containerColor = DesignTheme.colors.background,
                         headlineColor = DesignTheme.colors.content,
                     ),
                     modifier = Modifier.clickable {
-                        onFiatSelect.invoke(currency)
-                    },
-                    headlineContent = {
-                        Row {
-                            Text(
-                                modifier = Modifier.padding(tinyPad.dp),
-                                text = currency.name,
-                                style = MaterialTheme.typography.labelMedium
-                                    .copy(textAlign = TextAlign.Left)
-                            )
-                            Spacer(modifier = Modifier.weight(1f))
-                            Text(
-                                modifier = Modifier.padding(tinyPad.dp),
-                                text = "${currency.code} (${currency.symbol})",
-                                style = MaterialTheme.typography.labelMedium
-                                    .copy(textAlign = TextAlign.Left)
-                            )
+                        if (language.code.isNotBlank()) {
+                            onLanguageSelect.invoke(language)
                         }
                     },
+                    headlineContent = {
+                        Text(
+                            modifier = Modifier,
+                            text = language.title,
+                            style = MaterialTheme.typography.bodyMedium
+                                .copy(textAlign = TextAlign.Left)
+                        )
+                    },
                     trailingContent = {
-                        if (selectedCurrency.code == currency.code) {
+                        if (selectedLanguage.code == language.code) {
                             Icon(
                                 Icons.Default.CheckCircle,
                                 contentDescription = null,
