+++ app/src/main/java/com/brainwallet/ui/screens/settings/settingsrows/SettingRowItem.kt
@@ -1,4 +1,4 @@
-package com.brainwallet.ui.screens.main.composable.settingsrows
+package com.brainwallet.ui.screens.settings.settingsrows
 
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.foundation.background
