+++ app/src/main/java/com/brainwallet/ui/screens/settings/settingsrows/ThemeSettingRowItem.kt
@@ -1,4 +1,4 @@
-package com.brainwallet.ui.screens.main.composable.settingsrows
+package com.brainwallet.ui.screens.settings.settingsrows
 
 import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.width
